Updated the writer to be able to accurately calculate object ref sizes.
The writer now calculates everything possible on the first pass when building the object table. That way we can determine the maximum reference value in the offset table, which is ultimately the value needed to be addressable by reference integers.

The downside to this approach is that everything is built in memory before being written to the stream. This effectively doubles the memory usage of the writer. However, I can't fathom a different way to get accurate ref sizes.